requirements.txt
# AAOS Production Dependencies
# Generated for Phase 4 Production Deployment

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.13

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL driver for production

# Message Queue
redis==5.0.1

# WebSocket
websockets==12.0
aiohttp==3.9.1

# Async Support
asyncio-throttle==1.0.2

# Logging & Monitoring
python-json-logger==2.0.7
structlog==23.2.0

# Configuration
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# HTTP Client
httpx==0.25.2

# Testing (dev dependencies)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
